split config into multiple files and package them
