+++ allms/config/app.py
@@ -1,11 +1,10 @@
-from dataclasses import dataclass
 from pathlib import Path
 
 import tzlocal
 
 import allms.version as version
-from .utils.time import Time
-from .utils.logger import AppLogger
+from allms.utils.time import Time
+from allms.utils.logger import AppLogger
 
 
 class AppConfiguration:
@@ -54,7 +53,7 @@ class AppConfiguration:
     max_vote_duration_min: int = 10
 
     # Path of the resource directories and other files
-    __parent_dir: Path = Path(__file__).parent
+    __parent_dir: Path = Path(__file__).parent.parent
     __resource_dir_root: Path = __parent_dir / "res"
     __data_dir_root: Path = __parent_dir.parent / "data"
 
@@ -68,63 +67,3 @@ class AppConfiguration:
     # Logging configuration
     log_dir = __data_dir_root / "logs"
     logger = AppLogger(clock=clock, log_dir=log_dir)
-
-
-class StyleConfiguration:
-    """ Class holding constants for styling purposes """
-    class_border: str = "border"
-    class_border_highlight: str = "highlight-border"
-    class_modal_container: str = "modal-container"
-
-
-class ToastConfiguration:
-    """ Class holding constants for toasts """
-    type_information: str = "information"
-    type_warning: str = "warning"
-    type_error: str = "error"
-
-    timeout: float = 2.0  # How long (in seconds) a toast is displayed on the screen
-
-
-class BindingConfiguration:
-    """ Class holding the global hotkey bindings """
-    # Bindings for main screen
-    main_show_about: str = "ctrl+a"
-
-    # Bindings for modal screens
-    modal_close_screen: str = "ctrl+w"
-
-    # Bindings for new chat creation screen
-    new_chat_randomize_scenario: str = "ctrl+r"
-    new_chat_randomize_agent_persona: str = "ctrl+r"
-    new_chat_customize_agents: str = "ctrl+s"
-    new_chat_load_from_saved: str = "ctrl+l"
-
-    # Bindings for chat screen
-    chatroom_show_scenario: str = "f1"
-    chatroom_show_your_persona: str = "f2"
-    chatroom_show_all_persona: str = "f3"
-    chatroom_modify_msgs: str = "f4"
-    chatroom_start_vote: str = "f5"
-    chatroom_send_message: str = "enter"
-    chatroom_quit: str = "ctrl+w"
-
-    # Bindings for modify message screen
-    modify_msgs_mark_unmark_delete: str = "ctrl+x"
-
-
-@dataclass(frozen=True)
-class RunTimeConfiguration:
-    """ Configuration class holding constants from CLI and YAML config file """
-
-    ai_model: str
-    offline_model: bool
-    ai_reasoning_lvl: str
-    max_agent_count: int
-    default_agent_count: int
-    enable_rag: bool
-    show_thought_process: bool
-    show_suspects: bool
-    save_directory: str
-    ui_dev_mode: bool
-    skip_intro: bool
