+++ allms/config/style.py
@@ -1,0 +1,6 @@
+class StyleConfiguration:
+    """ Class holding constants for styling purposes """
+
+    class_border: str = "border"
+    class_border_highlight: str = "highlight-border"
+    class_modal_container: str = "modal-container"
